tz 15

+++ src/main/java/ru/practicum/shareit/item/dto/ItemResponseDto.java
@@ -1,10 +1,6 @@
 package ru.practicum.shareit.item.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.Builder;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,5 +12,5 @@ public class ItemResponseDto {
     private String name;
     private String description;
     private Boolean available;
-    private Integer requestId;
+    private Long requestId;
 }
